useIndicatorHeight, useWindowSize
