+++ src/components/community/Levels/components/AccessIndicator/index.ts
@@ -1,0 +1,3 @@
+import AccessIndecator from "./AccessIndicator"
+
+export default AccessIndecator
