+++ src/components/community/Levels/components/AccessIndicator/types.ts
@@ -1,0 +1,9 @@
+import { LevelIndicatorState } from "../Level/hooks/useLevelIndicatorState"
+
+type LevelState = {
+  isDisabled: boolean
+  element: HTMLElement
+  state?: LevelIndicatorState
+}
+
+export default LevelState
